# docs/swagger.yaml
basePath: /
definitions:
...
  dnsapi.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  dnsapi.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dnsapi.Record:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      prio:
        type: integer
      ttl:
        type: integer
      type:
        description: A, AAAA, CNAME, TXT, SRV
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
...
  dnsapi.UpdateZoneRequest:
    properties:
      abuse_email:
        example: admin@example.com
        type: string
      owner:
        example: alice
        type: string
      tags:
        example: tag1,tag2
        type: string
    type: object
  dnsapi.Zone:
    properties:
      abuse_email:
        type: string
      committed_serial:
        type: string
      created_at:
        type: string
      delete:
        type: boolean
      domain:
        type: string
      id:
        type: integer
      owner:
        type: string
      records:
        items:
          $ref: '#/definitions/dnsapi.Record'
        type: array
      serial:
        type: string
      tags:
        description: Tags separated by comma
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: Simple DNS API for managing Bind zones and records.
  title: DNS API
  version: "1.1"
paths:
  /bind/config:
    put:
      consumes:
      - text/plain
      description: Receives a rendered named.conf include file, writes it to disk,
        and reloads bind9.
      operationId: put-bind-config
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Write bind config (secondary mode)
      tags:
      - secondary
  /bind/refresh/{domain}:
    post:
      description: Forces a zone refresh using the configured refresh command.
      operationId: post-zone-refresh
      parameters:
      - description: Zone domain name
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Refresh zone (secondary mode)
      tags:
      - secondary
  /bind/reload:
    post:
      description: Reloads bind9 via the configured reload command.
      operationId: post-bind-reload
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Reload bind9 (secondary mode)
      tags:
      - secondary
  /bind/zones/{domain}:
    delete:
      description: Removes a zone file from the bind zone directory.
      operationId: delete-zone-file
      parameters:
      - description: Zone domain name
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete zone file (secondary mode)
      tags:
      - secondary
    put:
      consumes:
      - text/plain
      description: Receives a zone file and writes it to the bind zone directory.
      operationId: put-zone-file
      parameters:
      - description: Zone domain name
        in: path
        name: domain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Write zone file (secondary mode)
      tags:
      - secondary
...
  /zones/{zone_id}:
    delete:
      description: Delete a zone and all its records.
      operationId: delete-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.MessageResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete zone
      tags:
      - zones
    get:
      description: Return one zone including records.
      operationId: get-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get zone
      tags:
      - zones
    put:
      consumes:
      - application/json
      description: Update zone metadata (tags and abuse email only). Use the records
        endpoints to manage records.
      operationId: update-zone
      parameters:
      - description: Zone ID
        in: path
        name: zone_id
        required: true
        type: integer
      - description: Zone metadata to update
        in: body
        name: zone
        required: true
        schema:
          $ref: '#/definitions/dnsapi.UpdateZoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dnsapi.Zone'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dnsapi.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update zone
      tags:
      - zones
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"